Revert changes to main.zig
The example project is meant to work with the ESP32C3, which has an RMT
LED strip. This is essentially just a reimplementation of the "blink"
example in Zig.

Users who have a different board will need to adapt the project to their
needs, but this starter project is not meant to support all hardware.